Revert 513079a7ae4, making TheCPU.mem_base 32-bit again.
This wraps addresses at 4G, for compatibility with native DPMI behaviour.
I kept the sanity warning about the offsets being within 128 bytes, but
remade cr[5], because otherwise gcc warns about overflowing this array.

+++ src/emu-i386/simx86/interp.c
@@ -2104,7 +2104,7 @@ repag0:
 
 /*6c*/	case INSb: {
 			unsigned short a;
-			unsigned long rd;
+			unsigned int rd;
 			CODE_FLUSH();
 			a = rDX;
 			if (!test_ioperm(a)) goto not_permitted;
@@ -2222,7 +2222,7 @@ repag0:
 			rAL = port_real_inb(a);
 			PC += 2; } break;
 /*6d*/	case INSw: {
-			unsigned long rd;
+			unsigned int rd;
 			int dp;
 			CODE_FLUSH();
 			if (!test_ioperm(rDX)) goto not_permitted;
